created news list hook
