+++ src/modules/News/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './useGetNews';
+export * from './useNewsList';
